+++ src/card/cards/control_duration.rs
@@ -6,3 +6,5 @@ pub const ALADDIN: CardDefinitionId = CardDefinitionId(1472);
 pub const THRULL_CHAMPION: CardDefinitionId = CardDefinitionId(1473);
 pub const WILLOW_SATYR: CardDefinitionId = CardDefinitionId(1474);
 pub const RUBINIA_SOULSINGER: CardDefinitionId = CardDefinitionId(1475);
+pub const CONTROL_MAGIC: CardDefinitionId = CardDefinitionId(1802);
+pub const STEAL_ARTIFACT: CardDefinitionId = CardDefinitionId(1803);